fix(SERVER): fix summary categories and generalize Active/non-active status
