+++ ui/src/components/common/SummaryPageLayout/style.css
@@ -12,7 +12,7 @@
 }
 
 .summary-page-layout-collapsed-text {
-  color: #393A3D;
+  color: #393a3d;
   font-size: 0.8125rem;
   font-style: normal;
   font-weight: 600;
@@ -20,5 +20,5 @@
 
 .summary-page-layout-separator {
   width: 0.0625rem;
-  background: #DAE3E8;
+  background: #dae3e8;
 }
