+++ ui/src/components/pages/Pipeline/partials/PipelineISBStatus/index.tsx
@@ -1,7 +1,6 @@
 import React from "react";
 import Box from "@mui/material/Box";
 import {
-  GetISBType,
   IconsStatusMap,
   ISBStatusString,
   UNKNOWN,
@@ -10,7 +9,6 @@ import {
 import "./style.css";
 
 export function PipelineISBStatus({ isbData }) {
-  const isbType = GetISBType(isbData?.isbService?.spec) || UNKNOWN;
   const isbStatus = isbData?.isbService?.status?.phase || UNKNOWN;
   return (
     <Box
@@ -58,65 +56,6 @@ export function PipelineISBStatus({ isbData }) {
               <span className="pipeline-logo-text">
                 {ISBStatusString[isbData?.status]}
               </span>
-            </Box>
-          </Box>
-        </Box>
-        <Box sx={{ width: "fit-content", marginRight: "2em", flexGrow: 1 }}>
-          <span className="pipeline-status-title">ISB SERVICES SUMMARY</span>
-          <Box
-            sx={{
-              display: "flex",
-              flexDirection: "row",
-              marginTop: "0.3125rem",
-            }}
-          >
-            <Box
-              sx={{
-                display: "flex",
-                flexDirection: "column",
-                marginLeft: "0.3125rem",
-              }}
-            >
-              <div className="pipeline-summary-text">
-                <span className="pipeline-summary-subtitle">Name: </span>
-              </div>
-              <div className="pipeline-summary-text">
-                <span className="pipeline-summary-subtitle">Type: </span>
-              </div>
-            </Box>
-            <Box
-              sx={{
-                display: "flex",
-                flexDirection: "column",
-                marginLeft: "0.3125rem",
-              }}
-            >
-              <div className="isb-status-text">
-                <span>{isbData?.name}</span>
-              </div>
-              <div className="isb-status-text">
-                <span>{isbType}</span>
-              </div>
-            </Box>
-            <Box
-              sx={{
-                display: "flex",
-                flexDirection: "column",
-                marginLeft: "2rem",
-              }}
-            >
-              <div className="pipeline-summary-text">
-                <span className="pipeline-summary-subtitle">
-                  <div className="pipeline-summary-text">
-                    <span className="pipeline-summary-subtitle">Size: </span>
-                    <span>
-                      {isbType && isbData?.isbService?.spec[isbType]
-                        ? isbData?.isbService?.spec[isbType].replicas
-                        : UNKNOWN}
-                    </span>
-                  </div>
-                </span>
-              </div>
             </Box>
           </Box>
         </Box>
